Add Bootstrap compiler interview
Adds an interview that compiles bootstrap themes for admin and developers,
according to https://suffolklitlab.org/legal-tech-class/docs/practical-guide-docassemble/theming-docassemble/#creating-a-custom-theme-from-source-instead-of-with-a-theme-generator.

Once per server, it installs bootstrap-5.2.3 in the `/tmp` dir, and then
on going through the interview, it will run the `css-compile` command in the
project, returning the standalone CSS file is there were no errors. The users
can either upload a SCSS file, or type one directly, if it's shorter, or they
want to iterate faster.

Also added to icon to the menu. There are better font-awesome icons (like a
bootstrap specific one), but docassemble won't render brand ones correctly,
so I'm sticking with a generic but unique `file-code` for now.