Show preview of what the theme will look like (#104)
* Show preview of what the theme will look like

Mostly pure HTML from https://bootswatch.com/default/, and using a trick from
https://stackoverflow.com/a/13883978/11416267 to add the generated theme to
the CSS of the page using Javascript. Works great, might want to consider
culling down some of the things shown to be just the docassemble components,
since there's a whole lot, and most of it, people won't have to worry about.

* Make files a bit clearer on page

I went ahead and bolded them, and put one of those horizontal lines between the files at the top and the examples below. Hopefully that makes it a bit clearer.