Review gen improvements (#113)
* Use display template to show the generated review

Mainly for the copy button. Also adds some custom CSS to keep the code font,
and let people resize it.

* Don't re-order block attrs

id and question should be at the top, etc.

* Skip ALDocument/Bundle, and DAStaticFile objects

Shouldn't be a part of the review screen.

* Use <strong> on review screen if multiline Q

Otherwise, will get a syntax error (mako doesn't `**` on the same lines as it's
keywords).

* Improvements to generated revisit screens

* use `continue button field` over `event`
* add `add_action()` by default
* allow people to edit things added to their tables

* Display currency on review screens

A lot nicer of a display for currency.

* Option to point all sections at review screen

Something I do by default in my interviews; it lets people go to the review
screen from any section, and see all of their answers, instead of splitting
it up by section.